# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Press Kits Service",
    "description": "API for generating, versioning, and publicly sharing company press kits",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://press-kits.mcpfactory.org"
    }
  ],
  "components": {
    "schemas": {
...
      "MediaKitStatus": {
        "type": "string",
        "enum": [
          "drafted",
          "generating",
          "validated",
          "denied",
          "failed",
          "archived"
        ],
        "description": "drafted = generation complete, pending review. generating = workflow in progress. validated = approved and live. denied = rejected by user. failed = generation error or timeout. archived = superseded by a newer version."
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Media kit not found"
          }
        },
        "required": [
          "error"
        ]
      },
      "MediaKitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "orgId": {
            "type": "string",
            "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
          },
          "brandIds": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "a6b5fdad-b31d-4fa2-b34b-1cec4cb21ce5"
            ]
          },
          "campaignId": {
            "type": "string",
            "nullable": true,
            "example": "c7d8e9f0-a1b2-3456-cdef-789012345678"
          },
          "featureSlug": {
            "type": "string",
            "nullable": true,
            "example": "press-kit-v2"
          },
          "workflowSlug": {
            "type": "string",
            "nullable": true,
            "example": "generate-press-kit"
          },
          "shareToken": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Public share token. Use with GET /public/{token} to access the kit without authentication.",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "title": {
            "type": "string",
            "nullable": true,
            "example": "Acme Corp Press Kit — Q1 2026"
          },
          "iconUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://cdn.example.com/brands/acme/icon.png"
          },
          "brandDomain": {
            "type": "string",
            "nullable": true,
            "description": "Brand domain used for logo resolution via logo.dev.",
            "example": "acme.com"
          },
          "mdxPageContent": {
            "type": "string",
            "nullable": true,
            "description": "Full MDX content of the press kit page.",
            "example": "# Acme Corp\n\nAcme Corp is a leading provider of innovative SaaS solutions..."
          },
          "parentMediaKitId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "ID of the kit this version was forked from. Forms a linked list of versions. Null for the first kit in a scope.",
            "example": null
          },
          "status": {
            "$ref": "#/components/schemas/MediaKitStatus"
          },
          "denialReason": {
            "type": "string",
            "nullable": true,
            "description": "Reason for denial (when status is 'denied') or failure message (when status is 'failed').",
            "example": null
          },
          "createdAt": {
            "type": "string",
            "example": "2026-03-29T10:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-03-29T10:05:00.000Z"
          },
          "publicUrl": {
            "type": "string",
            "nullable": true,
            "description": "Full public URL for the press kit page. Ready to open in a browser.",
            "example": "https://press-kits.distribute.you/public/f47ac10b-58cc-4372-a567-0e02b2c3d479"
          }
        },
        "required": [
          "id",
          "orgId",
          "brandIds",
          "campaignId",
          "featureSlug",
          "workflowSlug",
          "shareToken",
          "title",
          "iconUrl",
          "brandDomain",
          "mdxPageContent",
          "parentMediaKitId",
          "status",
          "denialReason",
          "createdAt",
          "updatedAt",
          "publicUrl"
        ]
      },
...
      "GenerationDataResponse": {
        "type": "object",
        "properties": {
          "currentKit": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaKitResponse"
              },
              {
                "nullable": true
              }
            ]
          },
          "instructions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "instruction": {
                  "type": "string"
                },
                "instructionType": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "instruction",
                "instructionType",
                "createdAt"
              ]
            }
          },
          "feedbacks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "denialReason": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "denialReason"
              ]
            }
          }
        },
        "required": [
          "currentKit",
          "instructions",
          "feedbacks"
        ]
      },
      "UpsertGenerationResultRequest": {
        "type": "object",
        "properties": {
          "mediaKitId": {
            "type": "string",
            "format": "uuid",
            "description": "Target kit ID. If omitted, finds the generating kit by orgId.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "orgId": {
            "type": "string",
            "description": "Org ID fallback when mediaKitId is not provided.",
            "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
          },
          "mdxContent": {
            "type": "string",
            "example": "# Generated Press Kit\n\nContent produced by the generation workflow..."
          },
          "title": {
            "type": "string",
            "example": "Acme Corp Press Kit — Q1 2026"
          },
          "iconUrl": {
            "type": "string",
            "example": "https://cdn.example.com/brands/acme/icon.png"
          }
        },
        "required": [
          "mdxContent"
        ]
      },
      "EmailDataResponse": {
        "type": "object",
        "properties": {
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MediaKitStatus"
              },
              {
                "nullable": true,
                "example": "validated"
              }
            ]
          },
          "title": {
            "type": "string",
            "nullable": true,
            "example": "Acme Corp Press Kit — Q1 2026"
          },
          "pressKitUrl": {
            "type": "string",
            "nullable": true,
            "description": "Full public URL for the press kit page.",
            "example": "https://press-kits.distribute.you/public/f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "content": {
            "type": "string",
            "nullable": true,
            "example": "# Acme Corp\n\nLeading SaaS provider..."
          }
        },
        "required": [
          "status",
          "title",
          "pressKitUrl",
          "content"
        ]
      }
    },
...
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "paths": {
...
    "/internal/media-kits/current": {
      "get": {
        "summary": "Get latest media kit for scope",
        "description": "Returns the most recently updated media kit for the given org, optionally scoped by brand and campaign. Used by other services to check current press kit state.",
        "tags": [
          "Internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Filter by brand UUID"
            },
            "required": false,
            "name": "brand_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Filter by campaign UUID"
            },
            "required": false,
            "name": "campaign_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Media kit or null if none exists",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/MediaKitResponse"
                    },
                    {
                      "nullable": true
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/internal/media-kits/generation-data": {
      "get": {
        "summary": "Get data for generation workflow",
        "description": "Returns the currently generating kit with its instructions and past denial feedbacks. Used by the generation workflow to build the LLM prompt.",
        "tags": [
          "Internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Target a specific kit. If omitted, finds the generating kit for the org."
            },
            "required": false,
            "name": "media_kit_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Generation data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationDataResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/media-kits/generation-result": {
      "post": {
        "summary": "Upsert generation result (workflow callback)",
        "description": "Called by the generation workflow on success. Updates the generating kit with the produced MDX content and transitions it to 'drafted' status.",
        "tags": [
          "Internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertGenerationResultRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Kit updated to drafted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaKitResponse"
                }
              }
            }
          },
          "404": {
            "description": "No generating kit found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/email-data/{orgId}": {
      "get": {
        "summary": "Get press kit data for email templates",
        "description": "Returns the validated press kit data for a given org. Used by transactional-email-service to populate email templates.",
        "tags": [
          "Internal"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "orgId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal org UUID from client-service",
              "example": "org_3ANNRtJtvq2vahygqOSJ7IjRfp1"
            },
            "required": true,
            "name": "x-org-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Internal user UUID from client-service",
              "example": "usr_7BXkm2pTq1wLnHyjRfK4d"
            },
            "required": true,
            "name": "x-user-id",
            "in": "header"
          },
          {
            "schema": {
              "type": "string",
              "description": "Run ID for tracing (from runs-service)",
              "example": "550e8400-e29b-41d4-a716-446655440000"
            },
            "required": true,
            "name": "x-run-id",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "Email data (all fields null if no validated kit exists)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailDataResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "apiKey": []
    }
  ]
}